refactor(audio-quote-shadow): set `audio.currentTime`, `audio.muted` manually
We set `audio.currentTime` and `audio.muted` manually for the following situation:
The default behavior is to autoplay the audio and quote shadow animation.

However, the modern browser will block autoplay if the audio
is not set to `muted`. (For more information, please see
https://developer.mozilla.org/en-US/docs/Web/Media/Autoplay_guide#autoplay_availability)
(aftermath: `audio.play()` will return promise rejection)

And the unlucky part is that we cannot set `muted` in the `audio` tag
due to this known issue (https://github.com/facebook/react/issues/10389).
(aftermath: we have to change `muted` attribute by manipulating DOM
element)

Because of the above issues, we cannot play the audio with sound only if
users click the audio button. At the same time, the quote animation works
well.
(aftermath: audio and shadow animation are not synced)

When the users click the audio button, we have to play the audio at the
right `currentTime` to catch up the shadow animation.
(counting the `currentTime` by ourselves)

And this patch handles the above situation.