chore(root): update package.json. Update workspaces list
Remove `packages/embed-code-generator` pkg from `workspaces`.
The reason we remove it is because we want `packages/embed-code-generator`
to have its own `node_modules`. We don't want it to use the
symbolic link to other subpackages.

Since we might update other subpackages at the same time,
if the codes in other subpackages are not committed and not packaged.
embed-code-generator might mistakenly import these undergoing pkgs
while building the webpack bundles.